+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- [TD-2672] Read confidential value from concept instead of content
+
 ## [3.24.3] 2020-06-10
 
 ### Fixed